STAT-362: Better workaround for partition bug
After more testing, it's discovered that the bug only occurs when the
gs:// prefix is present. There is a path mismatch when reading the
inidvidual files from the directory. This occurs within the Arrow C code.

We handle this by simply removing the prefix before passing the path to
ParquetDataset.